+++ cpp/cpp/model/PolarCanvas.h
@@ -93,7 +93,7 @@ public:
     tr1::unordered_map<unsigned int,EngineShape*> fixedShapes;
     void tryNextEngineShape();
     void resetFixedShapes();
-    vector<int> fixShape(int sid, int x, int y, double rotation, double scaleX, double scaleY);
+    vector<int> fixShape(int sid, double x, double y, double rotation, double scaleX, double scaleY);
     vector<int> fixShape(int sid);
     double getSuccessRate();
     void registerShape(EngineShape* shape);
